fix https://github.com/pytorch/pytorch/issues/96406 and https://github.com/pytorch/pytorch/issues/98979

cc soumith voznesenskym penguinwu anijain2305 EikanWang jgong5 Guobing-Chen zhuhaozhe blzheng Xia-Weiwen wenzhe-nrv jiayisunx peterbell10 desertfire

[ghstack-poisoned]

+++ torch/_inductor/mkldnn.py
@@ -287,7 +287,7 @@ def pack_module(gm: torch.fx.GraphModule):
             if type(cur_module) in computation_op_packed_map:
                 if (
                     cur_module.weight.device != torch.device("cpu")
-                    or cur_module.weight.dtype not in [torch.float16, torch.float32]
+                    or cur_module.weight.dtype not in [torch.bfloat16, torch.float32]
                     or node.args[0].meta["val"].device != cur_module.weight.device
                     or node.args[0].meta["val"].dtype != cur_module.weight.dtype
                 ):
